requirements.txt
annotated-types==0.7.0
anyio==4.7.0
click==8.1.8
fastapi==0.115.6
graphql-core==3.2.5
h11==0.14.0
idna==3.10
libcst==1.5.1
markdown-it-py==3.0.0
mdurl==0.1.2
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
strawberry-graphql==0.256.0
typer==0.15.1
typing_extensions==4.12.2
uvicorn==0.34.0
